Migrate libc intrinsics to their own module
This is more sustainable as we add more intrinsics/overrides, e.g.
for Rust and libcxx
